pipeline/requirements.txt
# ARGO FloatChat - Complete Requirements
# Install with: pip install -r requirements.txt

# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
streamlit==1.28.1

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.23

# AI and ML
openai==1.3.7
groq==0.4.1
chromadb==0.4.18
sentence-transformers==2.2.2

# Data Processing
pandas==2.1.4
numpy==1.24.4
scipy==1.11.4

# Visualization
plotly==5.17.0
matplotlib==3.8.2
seaborn==0.13.0

# HTTP and API
httpx==0.25.2
requests==2.31.0

# Utilities
python-dotenv==1.0.0
pydantic==2.5.0
python-multipart==0.0.6

# Development
pytest==7.4.3
pytest-cov==4.1.0
black==23.11.0
flake8==6.1.0

# Optional: Ollama support
# ollama==0.1.7  # Uncomment if using Ollama
